creating tables

+++ task_2.sql
@@ -20,7 +20,7 @@ CREATE TABLE Customers (
 );
 
 CREATE TABLE Orders (
-    order_id BIGINT PRIMARY KEY,
+    order_id INT PRIMARY KEY,
     customer_id BIGINT,
     order_date DATE,
     FOREIGN KEY (customer_id) REFERENCES Customers (customer_id)
